First upload with basic scripts
The Plankton Muse scripts are the results of the process of composition of Jonas Hocherman's piece "Plankton Muse", for Big Band. This piece is based on the sonification of data representing the dynamic growth of phytoplankton in the Baltic Sea during the year 2024. The data was captured by the Finnish Environment Institute (SYKE) and shared with the author for artistic purposes.